requirements.txt
asgiref==3.4.1
boto3==1.20.35
botocore==1.23.35
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
drf-jwt==1.19.2
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
django-jazzmin==2.4.9
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.5
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
sendgrid==6.9.4
sentry-sdk==1.5.4
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.64.0
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
websocket-client==1.2.3
shortuuid==1.0.8
